recover Bits type to uint32 from uint64 in blockhead

+++ blockchain/blockindex.go
@@ -108,9 +108,9 @@ type blockNode struct {
 func initBlockNode(node *blockNode, blockHeader *wire.BlockHeader, parent *blockNode) {
 	*node = blockNode{
 		hash:       blockHeader.BlockHash(),
-		workSum:    CalcWork(uint32(blockHeader.Bits)),
+		workSum:    CalcWork(blockHeader.Bits),
 		version:    blockHeader.Version,
-		bits:       uint32(blockHeader.Bits),
+		bits:       blockHeader.Bits,
 		nonce:      uint32(blockHeader.Nonce),
 		timestamp:  blockHeader.Timestamp.Unix(),
 		merkleRoot: blockHeader.MerkleRoot,
@@ -145,7 +145,7 @@ func (node *blockNode) Header() wire.BlockHeader {
 		PrevBlock:  *prevHash,
 		MerkleRoot: node.merkleRoot,
 		Timestamp:  time.Unix(node.timestamp, 0),
-		Bits:       uint64(node.bits),
+		Bits:       node.bits,
 		Nonce:      uint64(node.nonce),
 	}
 }
