+++ wire/msgheaders_test.go
@@ -62,7 +62,7 @@ func TestHeaders(t *testing.T) {
 func TestHeadersWire(t *testing.T) {
 	hash := mainNetGenesisHash
 	merkleHash := blockOne.Header.MerkleRoot
-	bits := uint64(0x1d00ffff)
+	bits := 0x1d00ffff
 	nonce := uint64(0x9962e301)
 	bh := NewBlockHeader(1, &hash, &merkleHash, bits, nonce)
 	bh.Version = blockOne.Header.Version
@@ -230,7 +230,7 @@ func TestHeadersWireErrors(t *testing.T) {
 
 	hash := mainNetGenesisHash
 	merkleHash := blockOne.Header.MerkleRoot
-	bits := uint64(0x1d00ffff)
+	bits := 0x1d00ffff
 	nonce := uint64(0x9962e301)
 	bh := NewBlockHeader(1, &hash, &merkleHash, bits, nonce)
 	bh.Version = blockOne.Header.Version
